+++ R/constants.R
@@ -14,53 +14,6 @@ MAX_PROGRESS_VALUE <- 100
 
 DROPDOWN_MENU_ICONS <- list(messages = "mail", notifications = "warning sign", tasks = "tasks")
 
-sidebar_js <- "
-  /* Code below is needed to trigger visibility on reactive Shiny outputs. */
-  /* Thanks to that users do not have to set suspendWhenHidden to FALSE.   */
-  var previous_tab;
-  $('#uisidebar .item').tab({
-    onVisible: function(target) {
-      if (previous_tab) {
-        $(this).trigger('hidden');
-      }
-      $(window).resize();
-      $(this).trigger('shown');
-      previous_tab = this;
-    }
-  });
-
-  $('#uisidebar')
-    .sidebar({
-      context: $('.pusher'),
-      transition: 'push',
-      dimPage: false,
-      closable: false
-    })
-    .sidebar('attach events', '#toggle_menu');
-"
-
-body_js <- shiny::HTML("
-  <script>
-    var set_pusher_size = function(on_click) {
-      var sidebar = $('#uisidebar');
-      var window_width = $(window).width();
-      var pusher = $('.container > .pusher');
-      var is_sidebar_visible_after_transition = sidebar.hasClass('visible')
-      if (on_click) {
-        is_sidebar_visible_after_transition = !is_sidebar_visible_after_transition
-      };
-      if (is_sidebar_visible_after_transition) {sidebar_width = sidebar.width()} else {sidebar_width = 0};
-      if (sidebar.hasClass('left')) {margin_left = sidebar_width} else {margin_left = 0};
-      if (sidebar.hasClass('right')) {margin_right = sidebar_width} else {margin_right = 0};
-      pusher.width(window_width - sidebar_width);
-      pusher.css({'margin-left' : margin_left, 'margin-right' : margin_right});
-    };
-    $('#toggle_menu').click(function() {set_pusher_size(true)});
-    $(document).ready(function() {set_pusher_size(false)});
-    $(window).resize(function() {set_pusher_size(false)});
-  </script>
-")
-
 dropdown_menu_js <- "$('.ui.dropdown').dropdown();"
 
 progress_bar_js <- "$('.progress').progress();"
@@ -80,7 +33,7 @@ light_semantic_palette <- c("#DC73FF", "#2ECC40", "#54C8FF", "#FF695E", "#FF851B
                             "#D9E778", "#6DFFFF", "#A291FB", "#FF8EDF", "#D67C1C", "#DCDDDE",
                             "#545454")
 
-names(light_semantic_palette) <- c("lightPurple" , "lightGreen", "lightBlue", "lightRed",
+names(light_semantic_palette) <- c("lightPurple", "lightGreen", "lightBlue", "lightRed",
                                    "lightOrange", "lightYellow", "lightOlive", "lightTeal",
                                    "lightViolet", "lightPink", "lightBrown", "lightGrey",
                                    "lightBlack")