+++ examples/minimal/app_no_sidebar.R
@@ -3,10 +3,7 @@ library(semantic.dashboard)
 
 ui <- dashboardPage(
   dashboardHeader(color = "blue"),
-  dashboardSidebar(side = "left", size = "thin", color = "teal", disable = TRUE,
-                   sidebarMenu(
-                     menuItem(tabName = "tab1", "Tab 1"),
-                     menuItem(tabName = "tab2", "Tab 2"))),
+  dashboardSidebar(),
   dashboardBody(tabItems(
     tabItem(tabName = "tab1", p("Tab 1")),
     tabItem(tabName = "tab2", p("Tab 2"))))